Implement url parameter-based language and data selection
This commit introduces the use of url parameters for language selection and specific data fetching on mount. The `onMounted` lifecycle hook is leveraged to get and set 'lang' and 'data' parameters from the url upon component mounting. Additionally, it updates 'i18n.js' file for including data in experience objects and fixes a couple of mistypes as well.